fjm-frontend/api/requirements.txt
Click==7.0
dnspython==1.16.0
eventlet==0.24.0
Flask==1.0.2
Flask-Cors==3.0.7
greenlet==0.4.15
itsdangerous==1.1.0
iso8601==0.1.12
Jinja2==2.10
MarkupSafe==1.1.0
monotonic==1.5
pytz==2018.7
PyYAML==3.13
six==1.12.0
Werkzeug==0.14.1
